Change Markdown button syntax
