+++ src/lib/components/molecules/Markdown.svelte
@@ -1,12 +1,10 @@
 <script lang="ts">
 	import {
-		ButtonTokenizerExtension,
 		MarkerHighlightTokenizerExtension,
 		SparklesHighlightTokenizerExtension
 	} from '$lib/utils/markdown';
 	import { marked } from 'marked';
 	import SvelteMarkdown from 'svelte-markdown';
-	import ButtonRenderer from '../markdown/ButtonRenderer.svelte';
 	import ComponentRenderer from '../markdown/ComponentRenderer.svelte';
 	import HeadingRenderer from '../markdown/HeadingRenderer.svelte';
 	import ImageRenderer from '../markdown/ImageRenderer.svelte';
@@ -18,11 +16,7 @@
 	export let content: string;
 
 	marked.use({
-		extensions: [
-			SparklesHighlightTokenizerExtension,
-			MarkerHighlightTokenizerExtension,
-			ButtonTokenizerExtension
-		]
+		extensions: [SparklesHighlightTokenizerExtension, MarkerHighlightTokenizerExtension]
 	});
 	const options = marked.defaults;
 
@@ -31,7 +25,6 @@
 		sparklesHighlight: SparklingHighlight,
 		markerHighlight: MarkerHighlight,
 		image: ImageRenderer,
-		button: ButtonRenderer,
 		blockquote: QuoteCalloutRenderer,
 		link: LinkRenderer,
 		heading: HeadingRenderer
